@@ -38,7 +38,7 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   google_fonts: ^6.2.1
-  board_datetime_picker: ^2.1.3
+  board_datetime_picker: ^2.1.4
   flutter_holo_date_picker: ^2.0.0
   multi_select_flutter: ^4.1.3
   shimmer: ^3.0.0
@@ -46,11 +46,12 @@ dependencies:
   http: ^1.2.2
   equatable: ^2.0.5
   fan_carousel_image_slider: ^1.0.1
-  font_awesome_flutter: ^10.7.0
+  font_awesome_flutter: ^10.8.0
   shared_preferences: ^2.3.2
   image_picker: ^1.1.2
   moyasar: ^2.0.13
-  path: ^1.9.0
+  flutter_rating_bar: ^4.0.1
+  webview_flutter: ^4.10.0
 
 dev_dependencies:
   flutter_test:
