@@ -1,12 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:makfy_new/Screens/CustomerOrdersPage.dart';
 import 'package:makfy_new/Screens/MyHomePage.dart';
+import 'package:makfy_new/Screens/PaymentPage.dart';
 import 'package:makfy_new/Screens/ProfilePage.dart';
 import 'package:makfy_new/Screens/ServicePage.dart';
 import 'package:makfy_new/Screens/ShoppingCertPage.dart';
 import 'package:makfy_new/Screens/createServicePage.dart';
 import 'package:makfy_new/Screens/loginPage.dart';
 import 'package:makfy_new/Screens/mainsectionPage.dart';
+import 'package:makfy_new/Screens/myDistrictsPage.dart';
 import 'package:makfy_new/Screens/registration_page.dart';
 import 'package:makfy_new/Screens/userServicesPage.dart';
 import 'package:makfy_new/Screens/myordersPage.dart';
@@ -26,15 +28,17 @@ class AppRoutes {
     '/login': (context) => LoginPage(),
     '/register': (context) => RegistrationPage(),
     '/profile': (context) => Profilepage(),
-    '/my_orders': (context) => Myorderspage(),
     '/main_section': (context) => Mainsectionpage(),
     '/sub_section': (context) => Subsectionpage(),
     '/service_page': (context) => ServicePage(),
     '/personal_profile': (context) => PersonalProfilePage(),
     '/shopping_cert': (context) => ShoppingCertPage(),
+    '/my_orders': (context) => ShoppingCertPage(),
     '/customer_orders': (context) => CustomerOrdersPage(),
     '/create_service': (context) => createServicePage(),
     '/user_page': (context) => userServicesPage(),
+    '/payment_page': (context) => PaymentPage(),
+    '/my_districts': (context) => MyDistrictsPage(),
   };
 }
 
@@ -50,6 +54,7 @@ class LoadingPage extends StatelessWidget {
   // دالة للتحقق من حالة تسجيل الدخول وتوجيه المستخدم
   void _checkLoginStatus(BuildContext context) async {
     bool isLoggedIn = await checkIfLoggedIn();
+
     if (isLoggedIn) {
       Navigator.pushReplacementNamed(
           context, '/home'); // إذا كان مسجلًا، انتقل إلى الصفحة الرئيسية
